Get string and numeric functions working correctly.

+++ src/ECMABasic55/Program.cs
@@ -48,14 +48,14 @@ namespace ECMABasic55
                 var source = Convert.ToString(args[0]);
                 var value = Convert.ToString(args[1]);
                 var index = Convert.ToInt32(args[2]);
-                return source.IndexOf(value, index - 1) + 1;
+                return (double)source.IndexOf(value, index - 1) + 1;
             });
 
             FunctionFactory.Instance.Define("POS", new[] { ExpressionType.String, ExpressionType.String }, args =>
             {
                 var source = Convert.ToString(args[0]);
                 var value = Convert.ToString(args[1]);
-                return source.IndexOf(value) + 1;
+                return (double)source.IndexOf(value) + 1;
             });
         }
